Create currents markers using L.divIcon with SVG graphic.
Currently experimenting with dynamically assigning an id to
the SVG data in each icon as we create the marker so I can
rotate them as needed. Maybe this is totaly hacky, but it's
surprisingly hard to rotate a marker in react-leaflet. All
the stackoverflow solutions are outdated and don't work with
the latest react-leaflet v3. I'm not yet clear on how to
convert the prior accepted solution (RotatedMarker component
modified to work w v2) to use hooks as v3 requires instead
of withLeaflet