Add react hooks to update marker rotation on click
Ultimately i won't use click to trigger the rotation, but
was an excuse to start using hooks and needed to test the
repeated rotation. Adding the useEffect hook broke my event
handler, so then figured out i needed the useEffect hook to
handle the side effect of rotating the marker. So the
assigning new state and executing the side effect basesd on
that change are separated into these diff hooks.