Rotate marker icon div via inline style, not SVG
No longer using the style markup in the SVG. That only partially worked.
Now specifying an inline style in the DivIcon's html, and it looks like we can rotate this directly this way. Can't use the css for this cuz each marker needs it's own independent rotation val. With the inline style we supply the rotation value via template string. This means your icon can be whatever you want, image or svg.